pyproject.toml
[project]
name = "python_template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.12",
     "biopython",
     "pytest"
]

[tool.setuptools.packages.find]
include = ["core*", "ui*", "input*", "export*"]
